+++ migrator/migrations/policymigrationhelper/policy_migrator_test.go
@@ -60,13 +60,10 @@ func comparePolicyWithDB(suite *policyMigratorTestSuite, bucket bolthelpers.Buck
 	suite.EqualValues(policy, &newPolicy)
 }
 
-func strPtr(s string) *string {
-	return &s
-}
-
 func testPolicy(id string) storage.Policy {
 	return storage.Policy{
 		Id:          id,
+		Name:        "name",
 		Remediation: "remediation",
 		Rationale:   "rationale",
 		Description: "description",
@@ -390,6 +387,7 @@ func (suite *policyMigratorTestSuite) TestStringFieldsAreUpdatedIfNecessary() {
 		policyID: {
 			FieldsToCompare: []FieldComparator{DescriptionComparator},
 			ToChange: PolicyUpdates{
+				Name:        strPtr("new name"),
 				Description: strPtr("new description"),
 				Rationale:   strPtr("new rationale"),
 				Remediation: strPtr("new remediation"),
@@ -404,12 +402,59 @@ func (suite *policyMigratorTestSuite) TestStringFieldsAreUpdatedIfNecessary() {
 	suite.NoError(insertPolicyIntoBucket(bucket, policyID, &policy))
 	suite.NoError(MigratePolicies(suite.db, policiesToMigrate, comparisonPolicies))
 
-	// Policy should've had description, rationale and remediation changed, but nothing else
+	// Policy should've had name, description, rationale and remediation changed, but nothing else
+	policy.Name = *policiesToMigrate[policyID].ToChange.Name
 	policy.Description = *policiesToMigrate[policyID].ToChange.Description
 	policy.Rationale = *policiesToMigrate[policyID].ToChange.Rationale
 	policy.Remediation = *policiesToMigrate[policyID].ToChange.Remediation
 
 	comparePolicyWithDB(suite, bucket, &policy)
+}
+
+// Test that the string fields are updated as necessary
+func (suite *policyMigratorTestSuite) TestPolicyIsEnabledOrDisabledIfNecessary() {
+	bucket := bolthelpers.TopLevelRef(suite.db, policyBucketName)
+	enabledPolicy := testPolicy("enabled-policy")
+	enabledPolicy.Disabled = true // start out disabled so we can test it getting enabled
+	disabledPolicy := testPolicy("disabled-policy")
+	disabledPolicy.Disabled = false // start out enabled so we can test it getting disabled
+	noChangePolicy := testPolicy("no-change-policy")
+	noChangePolicy.Disabled = true // start out disabled so we can test it not getting changed
+
+	policiesToMigrate := map[string]PolicyChanges{
+		enabledPolicy.Id: {
+			FieldsToCompare: []FieldComparator{DescriptionComparator},
+			ToChange:        PolicyUpdates{Disable: boolPtr(false)},
+		},
+		disabledPolicy.Id: {
+			FieldsToCompare: []FieldComparator{DescriptionComparator},
+			ToChange:        PolicyUpdates{Disable: boolPtr(true)},
+		},
+		noChangePolicy.Id: {
+			FieldsToCompare: []FieldComparator{DescriptionComparator},
+			ToChange:        PolicyUpdates{Disable: nil},
+		},
+	}
+
+	comparisonPolicies := map[string]*storage.Policy{
+		enabledPolicy.Id:  &enabledPolicy,
+		disabledPolicy.Id: &disabledPolicy,
+		noChangePolicy.Id: &noChangePolicy,
+	}
+
+	for id, policy := range comparisonPolicies {
+		suite.NoError(insertPolicyIntoBucket(bucket, id, policy))
+	}
+	suite.NoError(MigratePolicies(suite.db, policiesToMigrate, comparisonPolicies))
+
+	// set the existing policies to what its state should be after migration
+	enabledPolicy.Disabled = false
+	disabledPolicy.Disabled = true
+	// noChangePolicy should remain the state it was originally
+
+	for _, policy := range comparisonPolicies {
+		comparePolicyWithDB(suite, bucket, policy)
+	}
 }
 
 // Test that policy section property is updated if asked
@@ -452,10 +497,10 @@ func (suite *policyMigratorTestSuite) TestPolicySectionIsUpdatedIfNecessary() {
 func (suite *policyMigratorTestSuite) TestPolicyIsUpdatedOnlyIfStringFieldComparisonsPass() {
 	bucket := bolthelpers.TopLevelRef(suite.db, policyBucketName)
 
-	policiesToTest := make([]*storage.Policy, 3)
+	policiesToTest := make([]*storage.Policy, 4)
 	comparisonPolicies := make(map[string]*storage.Policy)
 
-	for i := 0; i < 3; i++ {
+	for i := 0; i < 4; i++ {
 		policy := testPolicy(fmt.Sprintf("policy%d", i))
 		policiesToTest[i] = &policy
 		suite.NoError(insertPolicyIntoBucket(bucket, policy.Id, &policy))
@@ -463,21 +508,31 @@ func (suite *policyMigratorTestSuite) TestPolicyIsUpdatedOnlyIfStringFieldCompar
 
 	comparisonPolicy0 := testPolicy("policy0")
 	// Everything but description should _not_ match
+	comparisonPolicy0.Name = "alt name"
 	comparisonPolicy0.Remediation = "alt remediation"
 	comparisonPolicy0.Rationale = "alt rationale"
 	comparisonPolicies["policy0"] = &comparisonPolicy0
 
 	comparisonPolicy1 := testPolicy("policy1")
 	// Everything but remediation should _not_ match
+	comparisonPolicy1.Name = "alt name"
 	comparisonPolicy1.Description = "alt desc"
 	comparisonPolicy1.Rationale = "alt rationale"
 	comparisonPolicies["policy1"] = &comparisonPolicy1
 
 	comparisonPolicy2 := testPolicy("policy2")
 	// Everything but rationale should _not_ match
+	comparisonPolicy2.Name = "alt name"
 	comparisonPolicy2.Description = "alt desc"
 	comparisonPolicy2.Remediation = "alt remediation"
 	comparisonPolicies["policy2"] = &comparisonPolicy2
+
+	comparisonPolicy3 := testPolicy("policy3")
+	// Everything but name should _not_ match
+	comparisonPolicy3.Description = "alt desc"
+	comparisonPolicy3.Remediation = "alt remediation"
+	comparisonPolicy3.Rationale = "alt rationale"
+	comparisonPolicies["policy3"] = &comparisonPolicy3
 
 	policiesToMigrate := map[string]PolicyChanges{
 		"policy0": {
@@ -492,11 +547,16 @@ func (suite *policyMigratorTestSuite) TestPolicyIsUpdatedOnlyIfStringFieldCompar
 			FieldsToCompare: []FieldComparator{RationaleComparator},
 			ToChange:        PolicyUpdates{Rationale: strPtr("policy2 new rationale")},
 		},
+		"policy3": {
+			FieldsToCompare: []FieldComparator{NameComparator},
+			ToChange:        PolicyUpdates{Name: strPtr("policy3 new name")},
+		},
 	}
 
 	policiesToTest[0].Description = *policiesToMigrate["policy0"].ToChange.Description
 	policiesToTest[1].Remediation = *policiesToMigrate["policy1"].ToChange.Remediation
 	policiesToTest[2].Rationale = *policiesToMigrate["policy2"].ToChange.Rationale
+	policiesToTest[3].Name = *policiesToMigrate["policy3"].ToChange.Name
 
 	suite.NoError(MigratePolicies(suite.db, policiesToMigrate, comparisonPolicies))
 
@@ -505,15 +565,64 @@ func (suite *policyMigratorTestSuite) TestPolicyIsUpdatedOnlyIfStringFieldCompar
 	}
 }
 
+func (suite *policyMigratorTestSuite) TestPolicyTrimsWhitespaceForStringComparisons() {
+	bucket := bolthelpers.TopLevelRef(suite.db, policyBucketName)
+
+	policiesToTest := make([]*storage.Policy, 4)
+	comparisonPolicies := make(map[string]*storage.Policy)
+	for i := 0; i < 4; i++ {
+		policy := testPolicy(fmt.Sprintf("policy%d", i))
+
+		// Modify the policy so that the string fields have leading and/or trailing spaces but are otherwise equal
+		policy.Name = fmt.Sprintf("  \t %s", policy.Name)
+		policy.Description = fmt.Sprintf("%s  \t", policy.Description)
+		policy.Remediation = fmt.Sprintf("  \t %s  \t", policy.Remediation)
+		policy.Rationale = fmt.Sprintf(" %s\n", policy.Rationale)
+
+		policiesToTest[i] = &policy
+		suite.NoError(insertPolicyIntoBucket(bucket, policy.Id, &policy))
+
+		// Use the default policy for comparison
+		comparisonPolicy := testPolicy(policy.Id)
+		comparisonPolicies[policy.Id] = &comparisonPolicy
+	}
+
+	policiesToMigrate := make(map[string]PolicyChanges)
+	for _, policy := range policiesToTest {
+		policiesToMigrate[policy.GetId()] = PolicyChanges{
+			FieldsToCompare: []FieldComparator{NameComparator, DescriptionComparator, RemediationComparator, RationaleComparator},
+			ToChange: PolicyUpdates{
+				Name:        strPtr("new name"),
+				Description: strPtr("new description"),
+				Remediation: strPtr("new remediation"),
+				Rationale:   strPtr("new rationale"),
+			},
+		}
+	}
+
+	suite.NoError(MigratePolicies(suite.db, policiesToMigrate, comparisonPolicies))
+
+	for _, policy := range policiesToTest {
+		// Policy should've changed to the new values
+		policy.Name = "new name"
+		policy.Description = "new description"
+		policy.Remediation = "new remediation"
+		policy.Rationale = "new rationale"
+
+		comparePolicyWithDB(suite, bucket, policy)
+	}
+}
+
 func (suite *policyMigratorTestSuite) TestPolicyIsNotUpdatedIfStringFieldComparisonsFail() {
 	bucket := bolthelpers.TopLevelRef(suite.db, policyBucketName)
 
-	policiesToTest := make([]*storage.Policy, 3)
+	policiesToTest := make([]*storage.Policy, 4)
 	comparisonPolicies := make(map[string]*storage.Policy)
-	for i := 0; i < 3; i++ {
+	for i := 0; i < 4; i++ {
 		policy := testPolicy(fmt.Sprintf("policy%d", i))
 
 		// Modify the policy so that they don't match the default ones
+		policy.Name = "alt name"
 		policy.Description = "alt desc"
 		policy.Remediation = "alt remediation"
 		policy.Rationale = "alt rationale"
@@ -538,6 +647,10 @@ func (suite *policyMigratorTestSuite) TestPolicyIsNotUpdatedIfStringFieldCompari
 		"policy2": {
 			FieldsToCompare: []FieldComparator{RationaleComparator},
 			ToChange:        PolicyUpdates{Rationale: strPtr("policy2 new rationale")},
+		},
+		"policy3": {
+			FieldsToCompare: []FieldComparator{NameComparator},
+			ToChange:        PolicyUpdates{Name: strPtr("policy3 new name")},
 		},
 	}